+++ cpp/include/cudf/groupby.hpp
@@ -67,11 +67,6 @@ struct aggregation_result {
   std::vector<std::unique_ptr<column>> results{};
 };
 
-struct groupby_groups {
-  std::unique_ptr<table> group_keys;
-  std::vector<size_type> group_offsets;
-  std::unique_ptr<table> group_values;
-};
 
 /**
  * @brief Groups values by keys and computes aggregations on those groups.
@@ -170,14 +165,36 @@ class groupby {
       std::vector<aggregation_request> const& requests,
       rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
 
-  groupby_groups groups(
-    rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource(),
-    cudaStream_t stream = 0);
+  /**
+   * @brief The grouped data corresponding to a groupby operation on a set of values.
+   *
+   * A `groups` object holds two tables of identical number of rows:
+   * a table of grouped keys and a table of grouped values. In addition, it holds
+   * a vector of integer offsets into the rows of the tables, such that
+   * `group_offsets[i+1] - group_offsets[i]` gives the size of group `i`.
+   */
+  struct groups {
+    std::unique_ptr<table> group_keys;
+    std::vector<size_type> group_offsets;
+    std::unique_ptr<table> group_values;
+  };
 
-  groupby_groups groups(
-    cudf::table_view values,
-    rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource(),
-    cudaStream_t stream = 0);
+  /**
+   * @brief Get the grouped keys and values corresponding to a groupby operation on a set of values
+   *
+   * Returns a `groups` object representing the grouped keys and values.
+   * If values is not provided, only a grouping of the keys is performed,
+   * and accessing the `group_values` will be `nullptr`.
+   *
+   * @param values Table representing values on which a groupby operation is to be performed
+   * @param mr Memory resource used to allocate the returned tables
+   * @return A `groups` object representing grouped keys and values
+   */
+  groups get_groups(
+      cudf::table_view values={},
+      rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource()
+  );
+
 
 private:
   table_view _keys;                    ///< Keys that determine grouping