+++ cpp/tests/groupby/sort/groupby_groups_test.cu
@@ -25,49 +25,44 @@
 namespace cudf {
 namespace test {
 
-template <typename K>
-struct groupby_group_keys_test : public cudf::test::BaseFixture {};
+struct groupby_group_keys_test : public BaseFixture {};
 
 template <typename V>
 struct groupby_group_keys_and_values_test : public cudf::test::BaseFixture {};
 
-using KeyTypes = cudf::test::Types<int8_t, int16_t, int32_t, int64_t, float, double>;
-
-TYPED_TEST_CASE(groupby_group_keys_test, KeyTypes);
 TYPED_TEST_CASE(groupby_group_keys_and_values_test, NumericTypes);
 
-TYPED_TEST(groupby_group_keys_test, basic)
+TEST_F(groupby_group_keys_test, basic)
 {
-  using K = TypeParam;
+  using K = int32_t;
 
   fixed_width_column_wrapper<K> keys {1, 1, 2, 1, 2, 3};
   fixed_width_column_wrapper<K> expect_group_keys {1, 1, 1, 2, 2, 3};
   std::vector<size_type> expect_group_offsets = {0, 3, 5, 6};
-  test_grouping_keys(keys, expect_group_keys, expect_group_offsets);
+  test_groups(keys, expect_group_keys, expect_group_offsets);
 }
 
-TYPED_TEST(groupby_group_keys_test, empty_keys)
+TEST_F(groupby_group_keys_test, empty_keys)
 {
-  using K = TypeParam;
+  using K = int32_t;
 
   fixed_width_column_wrapper<K> keys {};
   fixed_width_column_wrapper<K> expect_group_keys {};
   std::vector<size_type> expect_group_offsets = {0};
-  test_grouping_keys(keys, expect_group_keys, expect_group_offsets);
+  test_groups(keys, expect_group_keys, expect_group_offsets);
 }
 
-
-TYPED_TEST(groupby_group_keys_test, all_null_keys)
+TEST_F(groupby_group_keys_test, all_null_keys)
 {
-  using K = TypeParam;
+  using K = int32_t;
 
   fixed_width_column_wrapper<K> keys ({1, 1, 2, 3, 1, 2}, all_null() );
   fixed_width_column_wrapper<K> expect_group_keys {};
   std::vector<size_type> expect_group_offsets = {0};
-  test_grouping_keys(keys, expect_group_keys, expect_group_offsets);
+  test_groups(keys, expect_group_keys, expect_group_offsets);
 }
 
-TYPED_TEST(groupby_group_keys_and_values_test, basic)
+TYPED_TEST(groupby_group_keys_and_values_test, basic_with_values)
 {
   using K = int32_t;
   using V = TypeParam;
@@ -77,9 +72,22 @@ TYPED_TEST(groupby_group_keys_and_values_test, basic)
   fixed_width_column_wrapper<V> values ({0, 0, 1, 1, 2, 2});
   fixed_width_column_wrapper<V> expect_group_values {2, 2, 1, 1, 0, 0};
   std::vector<size_type> expect_group_offsets = {0, 1, 2, 3, 4, 5, 6};
-  test_grouping_keys_and_values(keys, values, expect_group_keys, expect_group_values, expect_group_offsets);
+  test_groups(keys, expect_group_keys, expect_group_offsets, values, expect_group_values);
 }
 
+TYPED_TEST(groupby_group_keys_and_values_test, some_nulls)
+{
+  using K = int32_t;
+  using V = TypeParam;
+
+  fixed_width_column_wrapper<K> keys ({1, 1, 3, 2, 1, 2},
+                                      {1, 0, 1, 0, 0, 1});
+  fixed_width_column_wrapper<K> expect_group_keys ({1, 2, 3}, all_valid());
+  fixed_width_column_wrapper<V> values ({1, 2, 3, 4, 5, 6});
+  fixed_width_column_wrapper<V> expect_group_values ({1, 6, 3});
+  std::vector<size_type> expect_group_offsets = {0, 1, 2, 3};
+  test_groups(keys, expect_group_keys, expect_group_offsets, values, expect_group_values);
+}
 
 } //namespace test
 } //namespace cudf